Create square_root.cpp
If x is perfect square it calculates the square root of a number else it calculates floor of square root of number in O(log n) time. Regular square root method calculates in O(√ n). So use binary search to calculate square root of number and reduce the time complexity to O(log n).